backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
openai==2.6.1
httpx==0.28.1
pydantic==2.11.9
python-dotenv==1.0.1
redis==5.0.1
redisvl==0.9.0
requests==2.32.5
